refactor: moved cocktailMusic service function into a seperate file
